+++ src/Sound.h
@@ -1,25 +1,25 @@
 #include <vector>
 #include <complex>
 
-int samplesFromTime(double time);
-double timeFromSamples(int samples);
+int samplesFromTime(float time);
+float timeFromSamples(int samples);
 
 class Sound{
 	private:
-		void lengthenVector(std::vector<std::vector<double> >& input, int start, int end, int nuvolength);
+		void lengthenVector(std::vector<std::vector<float> >& input, int start, int end, int nuvolength);
 	public:
 		int windowLength;
 		int hop;
 		int hops;
-		std::vector<double> rawSynthesize();
+		std::vector<float> rawSynthesize();
 
-		std::vector< std::vector<double> > magnitudes;
-		std::vector< std::vector<double> > frequencies;
+		std::vector< std::vector<float> > magnitudes;
+		std::vector< std::vector<float> > frequencies;
 
-		double duration;
+		float duration;
 		int sampleRate;
 
-		Sound(std::vector<double> pcm=std::vector<double>(0,0), int overlap=8, int windowSize=1024, int rate=44100);
-		Sound compatibleSound(std::vector<double> pcm);
-		std::vector<double> synthesize();
+		Sound(std::vector<float> pcm=std::vector<float>(0,0), int overlap=8, int windowSize=1024, int rate=44100);
+		Sound compatibleSound(std::vector<float> pcm);
+		std::vector<float> synthesize();
 };
